backend/artifact_keeper/requirements.txt
pylint 
jedi
fastapi==0.103.2
pydantic-settings==2.0.3
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
SQLAlchemy==2.0.21
alembic==1.12.0
httpx==0.25.0
pytest==7.4.2
email-validator==2.1.0
passlib[bcrypt]==1.7.4
fastapi-mail
Jinja2==3.1.2
PyJWT==2.8.0
python-multipart==0.0.6
psycopg2

requests
